emad,add sechedule run

+++ app/Http/Controllers/Api/HomeController.php
@@ -13,16 +13,24 @@ class HomeController extends Controller
 {
     public function msisdn(Request $request)
     {
+        $output = 0 ;
         $result = Activation :: where("msisdn",$request->msisdn)->where("serviceid",$request->serviceid)->orderBy("created_at", "desc")->first(['id','msisdn','serviceid']);
         if($request->msisdn == $result["msisdn"]){
             $sub    = Subscriber :: where("activation_id",$result->id)->first();
-            $unsub  = new Unsubscriber();
-            $unsub->activation_id = $sub->activation_id;
-            $unsub->save();
-            $sub->delete();
+            if($sub ){
+                $unsub  = new Unsubscriber();
+                $unsub->activation_id = $sub->activation_id;
+                $unsub->save();
+                $sub->delete();
+                $output = 1 ;
+            }else{
+                $output = 0 ;
+            }
+
         }else{
-            return false;
+            $output = 0 ;
         }
+        return $output ;
 
     }
 }
